third_party/blink/web_tests/platform/generic/editing/execCommand/reset-values-after-navigation-expected.txt

PASS queryCommandValues["DefaultParagraphSeparator"] is "div"
PASS queryCommandValues["StyleWithCSS"] is ""
PASS successfullyParsed is true

TEST COMPLETE
